+++ tiger-rbel/src/main/java/de/gematik/rbellogger/util/RbelPathAble.java
@@ -5,8 +5,12 @@
 package de.gematik.rbellogger.util;
 
 import de.gematik.rbellogger.data.RbelMultiMap;
+import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.stream.Collectors;
 
 /**
  * Enables the usage of the RbelPathExecutor. The methods are called by the RbelPathExecutor, but
@@ -42,5 +46,22 @@ public abstract class RbelPathAble {
   /** Should this element be present in the final RbelPath results? */
   public boolean shouldElementBeKeptInFinalResult() {
     return true;
+  }
+
+  public String findNodePath() {
+    LinkedList<Optional<String>> keyList = new LinkedList<>();
+    final AtomicReference<RbelPathAble> ptr = new AtomicReference<>(this);
+    while (ptr.get().getParentNode() != null) {
+      keyList.addFirst(
+        ptr.get().getParentNode().getChildNodesWithKey().stream()
+          .filter(entry -> entry.getValue() == ptr.get())
+          .map(Map.Entry::getKey)
+          .findFirst());
+      ptr.set(ptr.get().getParentNode());
+    }
+    return keyList.stream()
+      .filter(Optional::isPresent)
+      .map(Optional::get)
+      .collect(Collectors.joining("."));
   }
 }