+++ tiger-rbel/src/main/java/de/gematik/rbellogger/util/RbelPathExecutor.java
@@ -4,6 +4,7 @@
 
 package de.gematik.rbellogger.util;
 
+import com.google.common.annotations.VisibleForTesting;
 import de.gematik.rbellogger.RbelOptions;
 import de.gematik.rbellogger.data.RbelElement;
 import de.gematik.rbellogger.exceptions.RbelPathException;
@@ -12,7 +13,6 @@ import de.gematik.test.tiger.common.jexl.TigerJexlExecutor;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -36,6 +36,7 @@ public class RbelPathExecutor<T extends RbelPathAble> {
     return result;
   }
 
+  @VisibleForTesting
   public static List<String> splitRbelPathIntoKeys(String rbelPath) {
     final String[] split = rbelPath.substring(1).trim().split("\\.(?!(\\.|[^\\[]*]))");
     final ArrayList<String> keys = new ArrayList<>();
@@ -104,26 +105,23 @@ public class RbelPathExecutor<T extends RbelPathAble> {
               .distinct()
               .toList();
       if (candidates.isEmpty() && RbelOptions.isActivateRbelPathDebugging()) {
-        List<RbelElement> asRbelElements =
-            lastIterationCandidates.stream()
-                .filter(RbelElement.class::isInstance)
-                .map(r -> (RbelElement) r)
-                .toList();
         log.warn(
             "No more candidate-nodes in RbelPath execution! Last batch of candidates had {}"
                 + " elements: \n"
                 + " {}",
-            asRbelElements.size(),
-            asRbelElements.stream()
-                .map(el -> el.printTreeStructure(Integer.MAX_VALUE, true))
-                .collect(Collectors.joining("\n")));
+            lastIterationCandidates.size(),
+            getPathList(lastIterationCandidates));
       }
     }
 
     final List<T> resultList =
         candidates.stream().filter(RbelPathAble::shouldElementBeKeptInFinalResult).toList();
     if (RbelOptions.isActivateRbelPathDebugging()) {
-      log.info("Returning {} result elements for RbelPath {}", resultList.size(), rbelPath);
+      log.info(
+          "Returning {} result elements for RbelPath {} (Results are {})",
+          resultList.size(),
+          rbelPath,
+          getPathList(resultList));
     }
     return resultList;
   }
@@ -132,8 +130,9 @@ public class RbelPathExecutor<T extends RbelPathAble> {
     if (RbelOptions.isActivateRbelPathDebugging()
         && targetObject instanceof RbelElement asRbelElement) {
       log.info(
-          "Executing RBelPath {} into root-element (limited view to {} levels)\n{}",
+          "Executing RBelPath {} into element '{}' (limited view to {} levels):\n{}",
           rbelPath,
+          targetObject.findNodePath(),
           Math.max(RbelOptions.getRbelPathTreeViewMinimumDepth(), keys.size()),
           asRbelElement.printTreeStructure(
               Math.max(RbelOptions.getRbelPathTreeViewMinimumDepth(), keys.size()), false));
@@ -167,37 +166,54 @@ public class RbelPathExecutor<T extends RbelPathAble> {
     final String selectorPart = parts[0];
     List<? extends RbelPathAble> keySelectionResult =
         executeNonFunctionalExpression(selectorPart, content);
-    if (parts.length == 1) {
+    if (parts.length == 1 || keySelectionResult.isEmpty()) {
       return keySelectionResult;
     } else {
-      final String functionalPart = parts[1].substring(0, parts[1].length() - 1);
-      if (NumberUtils.isParsable(functionalPart)) {
-        final int selectionIndex = Integer.parseInt(parts[1].substring(0, parts[1].length() - 1));
-        if (keySelectionResult.size() <= selectionIndex) {
-          return Collections.emptyList();
-        }
-        return List.of(keySelectionResult.get(selectionIndex));
-      } else {
-        return keySelectionResult.stream()
-            .map(
-                candidate -> executeFunctionalExpression(functionalPart, candidate.getParentNode()))
-            .flatMap(List::stream)
-            .toList();
+      return filterResultsThroughFunctionalSelector(
+          keySelectionResult, parts[1].substring(0, parts[1].length() - 1), selectorPart.isEmpty());
+    }
+  }
+
+  private List<? extends RbelPathAble> filterResultsThroughFunctionalSelector(
+      List<? extends RbelPathAble> keySelectionResult,
+      String functionalPart,
+      boolean selectorPartIsEmpty) {
+    if (RbelOptions.isActivateRbelPathDebugging()) {
+      log.info(
+          "Filtering resulting nodes '{}' through functional expression '{}'",
+          getPathList(keySelectionResult),
+          functionalPart);
+    }
+
+    if (NumberUtils.isParsable(functionalPart)) {
+      final int selectionIndex = Integer.parseInt(functionalPart);
+      if (keySelectionResult.size() <= selectionIndex) {
+        return Collections.emptyList();
       }
+      return List.of(keySelectionResult.get(selectionIndex));
+    } else {
+      return keySelectionResult.stream()
+          .map(
+              candidate ->
+                  executeFunctionalExpression(functionalPart, candidate, selectorPartIsEmpty))
+          .flatMap(List::stream)
+          .toList();
     }
   }
 
   private List<? extends RbelPathAble> executeNonFunctionalExpression(
       String key, RbelPathAble content) {
-    if (key.isEmpty() || key.equals("*")) {
+    if (key.equals("*")) {
       return content.getChildNodes();
+    } else if (key.isEmpty()) {
+      return List.of(content);
     } else {
       return content.getAll(key);
     }
   }
 
   private List<? extends RbelPathAble> executeFunctionalExpression(
-      final String functionExpression, final RbelPathAble content) {
+      final String functionExpression, final RbelPathAble content, boolean selectorPartIsEmpty) {
     if (functionExpression.startsWith("'") && functionExpression.endsWith("'")) {
       return executeNamedSelection(functionExpression, content);
     } else if (functionExpression.equals("*")) {
@@ -205,7 +221,9 @@ public class RbelPathExecutor<T extends RbelPathAble> {
     } else if (functionExpression.startsWith("?")) {
       if (functionExpression.startsWith("?(") && functionExpression.endsWith(")")) {
         return findChildNodesByJexlExpression(
-            content, functionExpression.substring(2, functionExpression.length() - 1));
+            content,
+            functionExpression.substring(2, functionExpression.length() - 1),
+            selectorPartIsEmpty);
       } else {
         throw new RbelPathException(
             "Invalid JEXL-Expression encountered (Does not start with '?(' and end with ')'): "
@@ -217,18 +235,27 @@ public class RbelPathExecutor<T extends RbelPathAble> {
   }
 
   private List<? extends RbelPathAble> findChildNodesByJexlExpression(
-      final RbelPathAble content, final String jexl) {
-    return content.getChildNodesWithKey().stream()
+      final RbelPathAble position, final String jexl, boolean selectorPartIsEmpty) {
+    List<RbelPathAble> candidates = new ArrayList<>();
+    if (selectorPartIsEmpty) {
+      candidates.addAll(position.getChildNodes());
+    } else {
+      candidates.add(position);
+    }
+    return candidates.stream()
         .parallel()
         .filter(
             candidate ->
                 TigerJexlExecutor.matchesAsJexlExpression(
                     jexl,
                     new TigerJexlContext()
-                        .withKey(candidate.getKey())
-                        .withCurrentElement(candidate.getValue())
+                        .withKey(candidate.getKey().orElse(null))
+                        .withCurrentElement(candidate)
                         .withRootElement(this.targetObject)))
-        .map(Map.Entry::getValue)
         .toList();
+  }
+
+  private static <T extends RbelPathAble> List<String> getPathList(List<T> resultList) {
+    return resultList.stream().map(RbelPathAble::findNodePath).map(path -> "$." + path).toList();
   }
 }