+++ tiger-rbel/src/test/java/de/gematik/rbellogger/util/RbelPathExecutorLoggingTest.java
@@ -112,7 +112,7 @@ class RbelPathExecutorLoggingTest {
   }
 
   @Test
-  void unsuccessfulRequestWithAmbiguousFinalPosition_expectTreeOfAllCandidates() {
+  void unsuccessfulRequestWithAmbiguousFinalPosition_expectListOfAllCandidates() {
     final ListAppender<ILoggingEvent> listAppender =
         listFollowingLoggingEventsForClass(RbelPathExecutor.class);
     jwtMessage.findRbelPathMembers("$.body.body.*.foobar");
@@ -125,7 +125,7 @@ class RbelPathExecutorLoggingTest {
     assertThat(
             listAppender.list.stream()
                 .map(ILoggingEvent::getFormattedMessage)
-                .map(str -> str.split("\\n\\n").length)
+                .map(str -> str.split(", ").length)
                 .max(Comparator.naturalOrder()))
         .get()
         .isEqualTo(34);